# WIP: Improve the efficiency of cupyx.scipy.ndimage.convolve (and correlate)