This PR rewrites the n-dimensional convolution kernel to make it more efficient. This is especially true when the filter itself only has a small number of non-zero entries.

Prior to this PR, in ``n`` dimensions, the kernel consisted of `n` nested for loops that looped over the filter kernel.

In this PR, the change is that the filter kernel is first converted to a 1d array containing the non-zero coefficients and a second array containing the coordinates of these coefficients. This allows replacing the ``n`` nested for loops by a **single** loop that only traverses the non-zero coefficients.

In my initial testing this seems to give a substantial performance benefit, even in the case when the filter does not have any non-zero entries. Here is a basic benchmark:

```Python
import cupy
from cupyx.scipy.ndimage import convolve

d = cupy.cuda.Device()
x = cupy.random.randn(256, 256, 256);

# densely populated filter
w = cupy.ones((5, 5, 5))
w /= w.size
%timeit convolve(x, w); d.synchronize()  

# sparsely populated filter
w2 = cupy.zeros((5, 5, 5))
w2[0, 0, 0] = 0.5
w2[-1, -1, -1] = -0.5
%timeit convolve(x, w2); d.synchronize() 
```

Prior to this PR the times for these dense and sparsely populated filters are equivalent. I measure:
```201 ms ± 7.45 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)```

After this PR, I get the following for the densely populated case:
```127 ms ± 283 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)```

and for the sparsely populated case:
```3.71 ms ± 69.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)```