@grlee77 2019-12-19
I can refactor slightly to pass `wshape` and the number of nonzero entries instead of the array `w` to `_generate_correlete_kernel`. That should allow restoring the caching of kernels.