@asi1024 2019-12-20
Why does this PR improve the performance of the densely populated inputs?