@asi1024 2019-12-20
> Prior to this PR the times for these dense and sparsely populated filters are equivalent. I measure:
> `201 ms ± 7.45 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

Some values are cached internally after the first function call. Please measure the time by repeating 10 times also in this case.