@asi1024 2019-12-20
I locally compared the performance in the densely populated case.
```py
import cupy
from cupyx.scipy.ndimage import convolve
from cupyx.time import repeat

d = cupy.cuda.Device()
x = cupy.random.randn(256, 256, 256);
w = cupy.random.randn(5, 5, 5)

perf = repeat(convolve, (x, w), n=50, n_warmup=10)
print(perf)
```

master(40b59e44d):
```
convolve            :    59.998 us   +/-54.305 (min:   46.976 / max:  339.244) us  72760.652 us   +/-79.005 (min:72704.094 / max:73062.813) us
```

This PR (bca5bba25):
```
convolve            :   359.820 us   +/-55.769 (min:  344.258 / max:  743.166) us  75956.443 us   +/-93.009 (min:75915.840 / max:76360.062) us
```