@grlee77 2019-12-20
I cannot do it immediately, but will take a look again at performance later today or tomorrow. I did not necessarily expect it to be faster for the densely populated case (but would also not expect it to be much slower).

It would also be possible to keep the current implementation and call it instead of the newly proposed one when all (or most) entries of the filter are non-zero.