@grlee77 2019-12-20
Okay, I reran with that `repeat` utility and also see a similar result to @asi1024. I think the CPU time being higher is to be expected due to the extra handling of weights to get the coordinates and values, but it looks like the GPU time is also slightly higher (212 vs 196 ms).

master(40b59e4):
```
convolve            :   234.678 us   +/-74.324 (min:  105.436 / max:  376.489) us  196379.369 us   +/-875.660 (min:194806.778 / max:197748.734) us
```
This PR (bca5bba):
```
convolve            :   759.932 us   +/-67.219 (min:  699.047 / max: 1095.894) us  213619.128 us   +/-916.371 (min:212165.634 / max:215121.918) us
```