@grlee77 2019-12-20
Given that there is some loss of performance for the fully dense case, I no longer want to merge this as-is. So I guess there are three main options:

1.) close the PR as is
2.) retain both kernel generation functions and select between the two either in an automated fashion (probably hard to do at least some minimal overhead) or via a new kwarg.
3.) improve the proposed solution so it is also faster in the dense case (not sure this is possible)

I am interested in how best to implement kernels that use a support mask of this type as that is a common pattern for a number of other functions in `scipy.ndimage` which use a `structure` or `footprint` argument. All of the morphological operations do this for example, so I don't think it is wasted effort even if we do not choose to use it here.

Have any of you worked on or have plans to work on these other types of image operations (e.g. `binary erosion`, `minimum_filter`, etc.)?