@grlee77 2019-12-20
I opened a separate PR with a simpler approach that I initially tried prior to this one: #2846.