@niboshi 2019-12-23
I feel it's a waste to just throw away this optimization, because it's siginificant in almost-zero case.
As an option for (3), how about doing this optimization only if `weights` is a sparse matrix?
That wouldn't work in `scipy.ndimage`, but I think it's at least better than introducing a new argument in (2).