@jakirkham 2020-03-06
Is this PR still needed @grlee77? Wasn't sure based on your earlier comment and the fact PR ( https://github.com/cupy/cupy/pull/2846 ) was merged.