@grlee77 2020-03-06
> Is this PR still needed @grlee77? Wasn't sure based on your earlier comment and the fact PR ( #2846 ) was merged.

I will go ahead and close it given that some of the benefit was obtained by the simpler change in #2846.  This implementation can have some benefit for filters that are mostly zeros, but I think that is a pretty niche case, and perhaps not worth maintaining a separate code path here.